refactor: change definition of VGA_Blit to use uint16_t and actually use width and height passed in

+++ remc2/portability/port_sdl_vga_mouse.h
@@ -27,7 +27,7 @@ typedef struct
 void VGA_Init(int width, int height, bool maintainAspectRatio);
 void VGA_Init(Uint32 flags, int width, int height, bool maintainAspectRatio);
 void VGA_Resize(int width, int height);
-void VGA_Blit(int width, int height, Uint8* srcBuffer);
+void VGA_Blit(uint16_t width, uint16_t height, Uint8* srcBuffer);
 void VGA_Debug_Blit(int width, int height, Uint8* srcBuffer);
 void VGA_Set_pallette(Uint8* pallettebuffer);
 void VGA_Set_pallette2(Uint8* pallettebuffer);
